+++ src/rpc/mod.rs
@@ -88,6 +88,7 @@ macro_rules! for_each_method {
         $callback!(crate::rpc::eth::EthGetTransactionByHash);
         $callback!(crate::rpc::eth::EthGetTransactionHashByCid);
         $callback!(crate::rpc::eth::EthCall);
+        $callback!(crate::rpc::eth::EthAddressToFilecoinAddress);
 
         // gas vertical
         $callback!(crate::rpc::gas::GasEstimateGasLimit);