+++ CHANGELOG.md
@@ -67,6 +67,9 @@
 - [#4613](https://github.com/ChainSafe/forest/issues/4613) Add support for the
   `Filecoin.EthCall` RPC method.
 
+- [#4686](https://github.com/ChainSafe/forest/issues/4686) Add support for the
+  `Filecoin.EthAddressToFilecoinAddress` RPC method.
+
 ### Changed
 
 - [#4583](https://github.com/ChainSafe/forest/pull/4583) Removed the expiration